refactor(email_api): split external_handlers en 4 sous-modules par domaine
Découpe external_handlers.rs (663 LOC) en:
- openapi.rs (~119 LOC): api_openapi_json, api_swagger_ui, api_external_openapi
- accounts.rs (~108 LOC): CRUD comptes externes + test IMAP
- folders.rs (~55 LOC): list/discover/mapping dossiers
- sync.rs (~385 LOC): handlers sync + messages + calendar helpers

mod.rs re-exporte tout via glob, aucun changement d'API.
